shutdown-timer
In this example, a System.Windows.Forms.Form object is created to display a numeric up-down control that allows the user to specify the value for the -t parameter. The form also contains OK and Cancel buttons to allow the user to confirm or cancel the operation.

The NumericUpDown control is created with a minimum value of 1 and a maximum value of 3600, and a default value of 30. When the OK button is clicked, the value of the NumericUpDown control is stored in the $tValue variable, and a second dialog box is displayed using the $host.UI.PromptForTime() method to allow the user to select the shutdown time.

If the user clicks the "OK" button in the second dialog box, the shutdown /s /t command is executed with the specified value for the -t parameter and the selected shutdown time.